moves snapshot token tests to pluggable suite
